+++ volumetrics/Voxelizer.h
@@ -23,8 +23,8 @@ namespace volumetrics {
 
 		void growSurfaceFromTriangle(FatVoxel *theVolume, unsigned int x, unsigned int y, unsigned int z,
 							 unsigned char triIndex, 
-							 vislib::Array<float> &surf, vislib::Array<BorderVoxel *> &border,
-                             float &surfSurf);
+							 vislib::Array<double> &surf, vislib::Array<BorderVoxel *> &border,
+                             double &surfSurf);
 
 		bool CellEmpty(FatVoxel *theVolume, unsigned x, unsigned y, unsigned z);
